CM-487: move picker to pickers folder
