+++ src/components/payroll/PayrollFilter.js
@@ -19,7 +19,7 @@ import {
   MODULE_NAME,
 } from '../../constants';
 import PayrollStatusPicker from './PayrollStatusPicker';
-import PaymentMethodPicker from './PaymentMethodPicker';
+import PaymentMethodPicker from '../../pickers/PaymentMethodPicker';
 
 const useStyles = makeStyles((theme) => ({
   form: {
